move search bar to its own file
